feat(management): move validators and filtering
